Removed createPersistentStore and extended createStore to handle persistent state
