+++ .changeset/long-kiwis-cheat.md
@@ -1,0 +1,5 @@
+---
+"dharma-core": minor
+---
+
+Removed createPersistentStore and extended createStore to handle persistent state
