+++ packages/dharma-core/src/index.ts
@@ -1,3 +1,2 @@
-export * from "./lib/createPersistentStore";
 export * from "./lib/createStore";
 export * from "./lib/merge";
