+++ packages/dharma-core/src/lib/createStore.node.test.ts
@@ -1,22 +1,23 @@
 // @vitest-environment node
 
 import { describe, expect, it, vi } from "vitest";
-import { createPersistentStore, StorageAPI } from "./createPersistentStore";
+import { createStore, StorageAPI } from "./createStore";
 
-describe("createPersistentStore", () => {
+describe("createStore (node)", () => {
   const baseConfig = {
+    persist: true,
     key: "test",
     initialState: { count: 0 },
     defineActions: vi.fn(),
   };
 
   it("should not throw in a node environment", () => {
-    expect(() => createPersistentStore(baseConfig)).not.toThrow();
+    expect(() => createStore(baseConfig)).not.toThrow();
   });
 
   it("should use the provided storage", () => {
     const storage = new CustomStorage();
-    const store = createPersistentStore({ ...baseConfig, storage });
+    const store = createStore({ ...baseConfig, storage });
     expect(storage.getItem("init_test")).toBe(JSON.stringify({ count: 0 }));
     store.set({ count: 1 });
     expect(storage.getItem("test")).toBe(JSON.stringify({ count: 1 }));
